Working (but minimal) gz3doom integration.
They share an EXE and an icon, but everything else is separate code. A different Main function (and thus complete program) is launched based on whether you have "-iwad" on the command line. If you choose Doom 1 or 2 from the BFG menu, it completely restarts the EXE with the specified command line.

I'll integrate it properly later, but for now it's usable.